Differring CLI output on push
My CLI output on cf push was pretty different from what you had - not sure why, but it was enough to confuse me (in finding the missing url) when following the directions. Might be due to the upgraded CLI (via #1), but it's a big change for a single digit version change.