Tentative timeline added
The idea is to have sense of the pace we need to work to get things done by the early October. 

Also I would like to raise the discussion about needs to have cloud account and access to chromebooks in advance so instructors can test their labs in exactly the same account they will be using during labs and set up the chromebooks the day before. This will eliminate issues with quotas etc.